factored out params
